+++ include/3dmath/MatrixOperations.h
@@ -10,30 +10,29 @@
 namespace math3d {
 
     // Convert a square matrix or the square matrix portion of an augmented matrix to an upper triangular matrix
-    template<typename DataType, size_t numRows, size_t numCols>
+    template<typename DataType, unsigned numRows, unsigned numCols>
     unsigned Matrix<DataType, numRows, numCols>::convertToUpperTriangular(Matrix<DataType, numRows, numCols>& upperTriangular) const {
 
         // Convert to upper triangular using Gaussian elimination with partial pivoting
         upperTriangular = *this;
-
         auto numRowSwaps = 0u;
 
         // NOTE 1: Rectangular matrices cannot be triangular. The loop assumes numRows = numColumns, but
         //         it uses numRows in the loop termination condition because we want to support turning
-        //         an augmented matrix, which is rectangular into a triangular matrix by ONLY considering
-        //         the non-augmented part, which is of dimension numRows x numRows
-        // NOTE 2: Last column skipped because there are no sub-diagonal elements in the last column to be
+        //         an augmented matrix, which is rectangular, into a triangular matrix by ONLY considering
+        //         the non-augmented part whose dimensions are numRows x numRows
+        // NOTE 2: Last column is skipped because there are no sub-diagonal elements in the last column to be
         //         turned into zeroes
-        constexpr size_t numberOfColumns = numRows - 1;
-        for (size_t columnIndex = 0; columnIndex < numberOfColumns; ++columnIndex) {
+        constexpr unsigned numberOfColumns = numRows - 1;
+        for (unsigned columnIndex = 0; columnIndex < numberOfColumns; ++columnIndex) {
 
             // Step 1: Pick a pivot element
             // This is the element with the largest absolute value in the current column
             // below the current row (which is also the current column since we are dealing with square matrices)
             Vector<DataType, numRows> column = upperTriangular[columnIndex];
-            size_t pivotIndex {columnIndex};
+            unsigned pivotIndex {columnIndex};
             DataType pivotValue = column[columnIndex];
-            for (size_t rowIndex = columnIndex + 1; rowIndex < numRows; ++rowIndex) {
+            for (unsigned rowIndex = columnIndex + 1; rowIndex < numRows; ++rowIndex) {
                 if (abs(column[rowIndex]) > abs(pivotValue)) {
                     pivotIndex = rowIndex;
                     pivotValue = column[pivotIndex];
@@ -55,7 +54,7 @@ namespace math3d {
             // a) multiplying the pivot row by matrix(rowIndex, columnIndex) / pivotValue
             // b) subtracting current row from the pivot row multiplied by the pivot value
             DataType factor {};
-            for (size_t rowIndex = columnIndex+1; rowIndex < numRows; ++rowIndex) {
+            for (unsigned rowIndex = columnIndex+1; rowIndex < numRows; ++rowIndex) {
                 factor = upperTriangular(rowIndex, columnIndex) / pivotValue;
                 upperTriangular.subtractRow(rowIndex, upperTriangular(pivotIndex) * factor);
             }
@@ -64,7 +63,7 @@ namespace math3d {
         return numRowSwaps;
     }
 
-    template<typename DataType, size_t numRows, size_t numCols>
+    template<typename DataType, unsigned numRows, unsigned numCols>
     DataType Matrix<DataType, numRows, numCols>::determinant() {
         static_assert(numRows == numCols, "Matrix::determinant(): Determinants are not defined for non-square matrices");
         static_assert(std::is_floating_point<DataType>::value, "Matrix::determinant(): Matrix's data type needs to be floating point");
@@ -77,14 +76,14 @@ namespace math3d {
         // Set the sign based on the number of row swaps
         DataType result {1.f};
         DataType const* upperTriangularData = upperTriangular;
-        for (size_t i = 0; i < numRows; ++i) {
+        for (unsigned i = 0; i < numRows; ++i) {
             result *= upperTriangularData[i * numRows + i];
         }
 
         return result * pow(-1, numRowSwaps);
     }
 
-    template<typename DataType, size_t numRows, size_t numCols>
+    template<typename DataType, unsigned numRows, unsigned numCols>
     Matrix<DataType, numRows, numCols> Matrix<DataType, numRows, numCols>::inverse() {
         static_assert(numRows == numCols, "Matrix::inverse(): Inverse is not defined. Only square matrices are invertible");
         static_assert(std::is_floating_point<DataType>::value,
@@ -102,7 +101,7 @@ namespace math3d {
         augmentedMatrix.convertToUpperTriangular(upperTriangular);
 
         // Step 3: Check pivot elements for zeroes
-        for (size_t rowIndex = 0; rowIndex < numRows; ++rowIndex) {
+        for (unsigned rowIndex = 0; rowIndex < numRows; ++rowIndex) {
             DataType pivotValue = upperTriangular.data[rowIndex * numRows + rowIndex];
             if (Utilities::isZero(pivotValue)) {
                 throw std::runtime_error("Matrix is not invertible");
@@ -114,8 +113,8 @@ namespace math3d {
         //         below the diagonal. Therefore, non-zero non-pivot elements are in columns [1, numRows)
         // NOTE 2: We use numRows instead of numCols for determining column index because we are dealing
         //         with a rectangular augmented matrix whose square portion is of size numRows x numRows
-        for (size_t columnIndex = numRows-1; columnIndex > 0; --columnIndex) {
-            for (size_t rowIndex = 0; rowIndex < numRows - 1; ++rowIndex) {
+        for (unsigned columnIndex = numRows-1; columnIndex > 0; --columnIndex) {
+            for (unsigned rowIndex = 0; rowIndex < numRows - 1; ++rowIndex) {
                 if (rowIndex >= columnIndex) {
                     continue;
                 }
@@ -129,9 +128,9 @@ namespace math3d {
         // Step 5: Divide each row by the pivot element
         // NOTE: Pivot elements are only present in the square matrix portion of the augmented matrix, which is of
         //       dimension numRows x numRows
-        for (size_t rowIndex = 0; rowIndex < numRows; ++rowIndex) {
+        for (unsigned rowIndex = 0; rowIndex < numRows; ++rowIndex) {
             auto pivotValue = upperTriangular.data[rowIndex * numRows + rowIndex];
-            for (size_t columnIndex = 0; columnIndex < 2 * numRows; ++columnIndex) {
+            for (unsigned columnIndex = 0; columnIndex < 2 * numRows; ++columnIndex) {
                 upperTriangular.data[columnIndex * numRows + rowIndex] /= pivotValue;
             }
         }
@@ -140,7 +139,7 @@ namespace math3d {
         return upperTriangular.template extract<numRows, numCols>(0, numCols);
     }
 
-    template<typename DataType, size_t numRows, size_t numCols>
+    template<typename DataType, unsigned numRows, unsigned numCols>
     Matrix<DataType, numRows, numCols> Matrix<DataType, numRows, numCols>::transpose() {
         Matrix <DataType, numCols, numRows> transposedMatrix;
         auto& transposedData = transposedMatrix.data;
@@ -153,13 +152,13 @@ namespace math3d {
     }
 
 
-    template<typename DataType, size_t numRows, size_t numCols>
-    void Matrix<DataType, numRows, numCols>::swapRows(size_t rowA, size_t rowB) {
+    template<typename DataType, unsigned numRows, unsigned numCols>
+    void Matrix<DataType, numRows, numCols>::swapRows(unsigned rowA, unsigned rowB) {
         if (rowA >= numRows || rowB >= numRows) {
             throw std::runtime_error("Matrix::swapRows(): Invalid row index encountered in swap call");
         }
         DataType tmp;
-        for (size_t i = 0; i < numCols; ++i) {
+        for (unsigned i = 0; i < numCols; ++i) {
             tmp = data[rowB];
             data[rowB] = data[rowA];
             data[rowA] = tmp;
@@ -168,18 +167,18 @@ namespace math3d {
         }
     }
 
-    template<typename DataType, size_t numRows, size_t numCols>
-    void Matrix<DataType, numRows, numCols>::addRow(size_t rowIndex, Vector<DataType, numCols> const& anotherRow) {
+    template<typename DataType, unsigned numRows, unsigned numCols>
+    void Matrix<DataType, numRows, numCols>::addRow(unsigned rowIndex, Vector<DataType, numCols> const& anotherRow) {
         if (rowIndex >= numRows) {
             throw std::runtime_error("Matrix::addRow(): Invalid row index!");
         }
-        for (size_t i = 0, index = rowIndex; i < numCols; ++i, index += numRows) {
+        for (unsigned i = 0, index = rowIndex; i < numCols; ++i, index += numRows) {
            data[index] += anotherRow[i];
         }
     }
 
-    template<typename DataType, size_t numRows, size_t numCols>
-    void Matrix<DataType, numRows, numCols>::subtractRow(size_t rowIndex, Vector<DataType, numCols> const& anotherRow) {
+    template<typename DataType, unsigned numRows, unsigned numCols>
+    void Matrix<DataType, numRows, numCols>::subtractRow(unsigned rowIndex, Vector<DataType, numCols> const& anotherRow) {
         addRow(rowIndex, static_cast<DataType>(-1)*anotherRow);
     }
 }
